Remove Grok into browser pages
