+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.2b2 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Remove Grok dependency for views.
+  [l34marr, rodfersou]
+
 - Use the `X-Robots-Tag` header to avoid indexing of image scales on default view;
   this will reduce the number of 404 (Not Found) responses generated by crawlers visiting the site in search of volatile content.
   [hvelarde]
