--- src/collective/cover/browser/contentchooser_templates/test_content_contentchooser.pt
@@ -1,18 +1,0 @@
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en"
-      xmlns:tal="http://xml.zope.org/namespaces/tal"
-      xmlns:metal="http://xml.zope.org/namespaces/metal"
-      xmlns:i18n="http://xml.zope.org/namespaces/i18n"
-      lang="en"
-      metal:use-macro="context/main_template/macros/master"
-      i18n:domain="collective.cover">
-<body>
-
-<metal:main fill-slot="main">
-    <tal:main-macro metal:define-macro="main">
-       <div class="content-contentchooser" tal:replace="structure context/@@select-content-contentchooser"></div>
-
-    </tal:main-macro>
-</metal:main>
-
-</body>
-</html>
