+++ src/collective/cover/browser/layoutpolicy.py
