+++ src/collective/cover/browser/templates/content_contentchooser.pt
