+++ src/collective/cover/browser/templates/pagelayout.pt
