+++ src/collective/cover/browser/templates/row.pt
