+++ src/collective/cover/browser/templates/search_list.pt
